Bump electron to 13.6.x
On Linux systems with glibc 2.34 (and newer), Chromium/Electron can crash with:

'GPU process isn't usable. Goodbye.'

A possible workaround is to launch with the app using the command line, and
passing '--disable-gpu-sandbox' as an argument, but that's not user-friendly.

This has been fixed in electron 13.x a long time ago:
https://github.com/electron/electron/commit/993ecb5bdd5c57024c8718ca6203a8f924d6d574